Added input devices Keyboard, Mouse, and Gamepad
Added the Keyboard, Mouse, and Gamepad classes for reading the input device state as it's fed through the input pipeline. These classes do not automaticaly query an underlying system for data and are instead written to via proxies; this allows the device classes to mainly act as views. These classes also bring with them the input enums Keyboard.Key, Mouse.Button, Gamepad.Button, and Gamepad.Axis as well as supporting constructs.