Added ButtonRule and AxisRule classes, supporting tests, and an updated test class for IntegratableInput's callbacks
Added ButtonRule and AxisRule classes to hold commands that will be bound to input. Maps with rules as values are the representation of "key bindings", in the case of the keyboard. The previously removed test class for IntegratableInput's callbacks have been re-added with a focus on expected exceptions while the callbacks' effects have been moved to the main test class for IntegratableInput.