Add Domain, GameSystem, and supporting tests and interfaces
Domains and GameSystems are a way to not only create mini-processes within a game but also hook into the processes in a managed way. This new extension point opens the door for subclasses to implement new game features at different levels of the framework from wide-ranging controls to player setup to more focused areas such as game rooms (which are likely to use many different systems).

- Add Domain for managing GameSystem subclass instances' lifecycles.
- Add GameSystem for client-created systems to hook into a larger execution path. This is the basis for implementing things like game weather or quick time events.
- Add SystemCoordinator, SystemGroupCoordinator, SystemDirectory, WritableSystemDirectory, and Prioritizable to help classes wrapping Domain for GameSystem support control exposed functionality
- Add DomainTest and GameSystemTest